requirements.txt
google-adk
python-dotenv
requests
pydantic
tenacity
beautifulsoup4
lxml
litellm
